GEO URL resolver

Construct and resolve URLs for GEO (Gene Expression Omnibus) resources. This function provides a unified interface for accessing GEO data either via NCBI’s accession-based query system or directly from GEO FTP/HTTPS servers. Depending on the accession type or requested famount, it automatically generates the correct URL.

Usage

geo_url(
  accession,
  famount = NULL,
  format = NULL,
  scope = NULL,
  over_https = NULL
)

Arguments

accession

A character of GEO accession IDs. Examples:

DataSets (GDS): "GDS505", "GDS606", "GDS1234", "GDS9999", etc.

Series (GSE): "GSE2", "GSE22", "GSE100", "GSE2000", etc.

Platforms (GPL): "GPL96", "GPL570", "GPL10558", etc.

Samples (GSM): "GSM12345", "GSM67890", "GSM112233", etc.

famount

A character specifying file/amount type requested. GEO data can be accessed through two sites:

Direct FTP/HTTPS file retrieval from GEO FTP server (file type):

"soft": SOFT (Simple Omnibus in Text Format) from GEO FTP site. When accession is DataSets or Series, this is the default.

"soft_full": full SOFT (Simple Omnibus in Text Format) files from GEO FTP site by DataSet (GDS) containging additionally contains up-to-date gene annotation for the DataSet Platform.

"miniml": MINiML (MIAME Notation in Markup Language, pronounced miniml) is an XML format that incorporates experimental data and metadata. MINiML is essentially an XML rendering of SOFT format.

"matrix": Series matrix file.

"annot": annotation files for Platforms.

"suppl": supplementary files.

The following table summarizes the compatibility between GEO accession types and file type options:

famount	GDS	GSE	GPL	GSM

SOFT (soft)	o	o	o	x

SOFTFULL (soft_full)	o	x	x	x

MINiML (miniml)	x	o	o	x

Matrix (matrix)	x	o	x	x

Annotation (annot)	x	x	o	x

Supplementaryfiles (suppl)	x	o	o	o

Accession-based queries to the NCBI GEO database (amount of data):

"none": Applicable only to DataSets; for DataSets, this is also the sole valid option.

"brief": accession attributes only.

"quick": accession attributes + first 20 rows of the data table.

"data": omits the accession's attributes, showing only links to other accessions and the full data table.

"full": accession attributes + complete data table.

format

A character specifying the output format (Only applicable to NCBI GEO database access):

"none": Applicable only to DataSets; for DataSets, this is also the sole valid option (no downloadable entry available).

txt/"text": machine-readable SOFT format (Simple Omnibus Format in Text).

"xml": XML format.

"html": human-readable format with hyperlinks (no downloadable entry available).

scope

A character specifying which GEO accessions to include (Only applicable to NCBI GEO database access).

"none": Applicable only to DataSets; for DataSets, this is also the sole valid option

"self": the queried accession only.

"gsm", "gpl", "gse": related samples, platforms, or series.

"all": all accessions related to the query (family view).

over_https

Logical scalar. If TRUE, connects to GEO FTP server via HTTPS (https://ftp.ncbi.nlm.nih.gov/geo); otherwise uses plain FTP (ftp://ftp.ncbi.nlm.nih.gov/geo). Only applicable to GEO FTP server access.

Value

A character of GEO URL.