Parse key-value pairs in GEO series soft file

Lots of GSEs now use "characteristics_ch*" meta header data for key-value pairs of annotation. If that is the case, this simply cleans the GSM GEOSoft @meta slot up and transforms the keys to column names and the values to column values.

Usage

parse_gsm_list(gsm_list)

Arguments

gsm_list

A list of GEOSoft, especially for @gsm slot in a GEOSeries object.

Value

a data.frame whose rows are samples and columns are the sample infos

Examples

gse201530_soft <- geo(
    "GSE201530",
    odir = tempdir(),
    gse_matrix = FALSE
)
#> Downloading 1 GSE full amount file from GEO Accession Site
#> Error in entity[[2L]]: subscript out of bounds
parse_gsm_list(gsm(gse201530_soft))
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'object' in selecting a method for function 'gsm': object 'gse201530_soft' not found