Search GEO database

Search the GDS database and return search results as a data.table.

Usage

geo_search(query, step = 500L, interval = NULL)

Arguments

query

A character string with the search term. The NCBI uses a fielded search syntax. For example, "Homo sapiens[ORGN]" searches the "Organism" field for Homo sapiens. See the GEO query tutorial for details. Searchable fields can be listed with rentrez::entrez_db_searchable("gds").

step

Integer. Number of records to fetch per request. Use a smaller value if requests fail.

interval

Numeric. Time interval (in seconds) between successive requests. Defaults to 0. Increase this value if requests fail due to rate limits.

Value

A data frame contains the search results

Details

The NCBI allows higher request limits (10 per second) when using an API key. You can set this key for the current R session with rentrez::set_entrez_key(), or permanently by setting the ENTREZ_KEY environment variable via Sys.setenv(). Once set, rentrez will automatically use this key for all NCBI requests. See the rentrez tutorial for details.

Examples

# Ensure you have an active internet connection before running the search.
# The `geo_search` function queries NCBI Entrez, which may have network
# restrictions and limited bandwidth usage for large queries. To avoid
# interruptions due tonetwork issues or rate limits, we wrap the call with
# try().
try(geo_search("diabetes[ALL] AND Homo sapiens[ORGN] AND GSE[ETYP]"))
#> Error : HTTP failure: 400
#> <?xml version="1.0" encoding="UTF-8" ?>
#> <!DOCTYPE eEfetchResult PUBLIC "-//NLM//DTD efetch 20131226//EN" "https://eutils.ncbi.nlm.nih.gov/eutils/dtd/20131226/efetch.dtd">
#> <eFetchResult>
#> 	<ERROR>Couldn&apos;t resolve #pmquerysrv-mz?dbaf=gds, the address table is empty. connection: #pmquerysrv-mz?dbaf=gdsservice: search</ERROR>
#> </eFetchResult>
#>