Search GEO database

Search the GDS database and return search results as a data.table.

Usage

geo_search(query, step = 500L, interval = NULL)

Arguments

query

A character string with the search term. The NCBI uses a fielded search syntax. For example, "Homo sapiens[ORGN]" searches the "Organism" field for Homo sapiens. See the GEO query tutorial for details. Searchable fields can be listed with rentrez::entrez_db_searchable("gds").

step

Integer. Number of records to fetch per request. Use a smaller value if requests fail.

interval

Numeric. Time interval (in seconds) between successive requests. Defaults to 0. Increase this value if requests fail due to rate limits.

Value

A data.table contains the search results

Details

The NCBI allows higher request limits (10 per second) when using an API key. You can set this key for the current R session with rentrez::set_entrez_key(), or permanently by setting the ENTREZ_KEY environment variable via Sys.setenv(). Once set, rentrez will automatically use this key for all NCBI requests. See the rentrez tutorial for details.

Examples

geo_search("diabetes[ALL] AND Homo sapiens[ORGN] AND GSE[ETYP]")
#> ■■■■■■■■■■                       500/1691 [324/s] | ETA:  4s
#> ■■■■■■■■■■■■■■■■■■■              1000/1691 [255/s] | ETA:  3s
#> ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■  1692/1691 [261/s] | ETA:  0s
#> → Parsing GEO records
#> ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■  1692/1691 [261/s] | ETA:  0s
#> Get records from NCBI for 1691 queries in 7.2s
#> 
#>                                                                                                                                                  Title
#>                                                                                                                                                 <char>
#>    1:                                               VAV3 antagonizes Skeletal Muscle Disulfidptosis to maintain Glucose Homeostasis in Type 2 Diabetes
#>    2: SmallRNA seq of leukocytes of chronic pain patients (CRPS, Nerve lesion and Diabetes polyneuropathy) and an SNI mice model ncRNA pain consortium
#>    3:                                        Coxsackievirus B infection invokes unique cell-type specific responses in primary human pancreatic islets
#>    4:                            Expression data from type 2 diabetes mellitus adipose-derived stem cells cultured with basic fibroblast growth factor
#>    5:                                            Engineered vasculature induces functional maturation of pluripotent stem cell-derived islet organoids
#>   ---                                                                                                                                                 
#> 1687:                                                        Gestational Diabetes Induces Placental Genes for Chronic Stress and Inflammatory Pathways
#> 1688:                                                                         laughter regulates postprandial blood glucose levels and gene expression
#> 1689:                                                                                                                             Diabetic nephropathy
#> 1690:                                                                                                    Muscle - atypical diabetes protein expression
#> 1691:                                                                                                           Type 2 diabetes and insulin resistance
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         Summary
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          <char>
#>    1:                                        The emergence and progression of type 2 diabetes mellitus (T2DM) are intricately linked to skeletal muscle loss, yet the exact mechanisms remain elusive. Here, our whole-exome and transcriptomic results indicate that VAV3 is mutated and down-regulated in T2DM patients associated with increased sulfur metabolism. Integrated analysis of human transcriptomics, mouse and cellular proteomics, glycol-metabolomics analyses suggest a potential association between VAV3, T2DM and disulfidptosis. more...
#>    2:                                                                                                                          Chronic neuropathic pain is widespread, but individual diagnostic biomarkers are unknown and primate specificity hampers rodent validation tests. Here, we report that microRNA (miR)-sequencing of blood leukocytes from 250 multi-centre pain patients revealed complex regional pain syndrome (CRPS)-related increases in the pain-related microRNA (miR)-21-5p and decreases in the cholinergic-targeted miR-335-5p. more...
#>    3:                                                                                                               Coxsackievirus B (CVB) infection has long been considered an environmental factor precipitating Type 1 diabetes (T1D), an autoimmune disease marked by loss of insulin-producing b cells within pancreatic islets. Previous studies have shown CVB infection negatively impacts islet function and viability but do not report on how virus infection individually affects the multiple cell types present in human primary islets. more...
#>    4:                  Diabetes affects ASCs characteristics such as: proliferation, differentiation and angiogenic capacity. MicroRNAs are able to target genes involved in vascular remodeling and promote or inhibit structural changes in the vessel wall. Adipose tissue-derived stem cells (ASCs) have the capacity to contribute to vascular remodeling. We used microarrays to detail the global miRNA expression profile underlying cell differentiation and identified distinct classes of up-regulated and down-regulated genes during this process.
#>    5: Blood vessels play a critical role in pancreatic islet function, yet current methods for deriving islet organoids from human pluripotent stem cells (SC-islets) lack vasculature. We engineered 3D vascularized SC-islet organoids by assembling SC-islet cells, human primary endothelial cells (ECs) and fibroblasts in a non-perfused model and a microfluidic device with perfused vessels. Vasculature improved stimulus-dependent Ca2+ influx into SC-β-cells, a hallmark of β-cell function that is blunted in non-vascularized SC-islets. more...
#>   ---                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          
#> 1687:                                    A physiological state of insulin resistance is required to preferentially direct maternal nutrients toward the feto-placental unit, allowing adequate growth of the fetus. When women develop gestational diabetes mellitus (GDM), insulin resistance is more severe and disrupts the intrauterine milieu, resulting in accelerated fetal development with increased risk of macrosomia. As a natural interface between mother and fetus, the placenta is the obligatory target of such environmental changes. more...
#> 1688:                                                                                                                                                                                                                                                                                 Sample tissue: peripheral blood Disease: diabetes Samples for gene expression analysis were obtained before the meal and 1.5 hours after the event. Event: listening to a Japanese comic story or a monotonous academic lecture without humor. Keywords: equivalent probe
#> 1689:                                                                                                                                                                                                                                                                                                                                                                                            Gene expression profiling in glomeruli from human kidneys with diabetic nephropathy Keywords = Diabetes Keywords = kidney Keywords = glomeruli Keywords: other
#> 1690:                                                                                                                                                                                                                                                                                                                                                                                                        Skeletal muscle biopsies from atypical diabetics at presentation and remission. Protein expression determined with antibody arrays Keywords: other
#> 1691:                                                                                                                                                                                    Global transcript profiling to identify differentially expressed skeletal muscle genes in insulin resistance, a major risk factor for Type II (non-insulin-dependent) diabetes mellitus.  Compared gene expression profiles of skeletal muscle tissues from 18 insulin-sensitive versus 17 insulin-resistant equally obese, non-diabetic Pima Indians. Keywords: other
#>                                Organism
#>                                  <char>
#>    1:                      Homo sapiens
#>    2:        Homo sapiens; Mus musculus
#>    3:                      Homo sapiens
#>    4: synthetic construct; Homo sapiens
#>    5:                      Homo sapiens
#>   ---                                  
#> 1687:                      Homo sapiens
#> 1688:                      Homo sapiens
#> 1689:                      Homo sapiens
#> 1690:                      Homo sapiens
#> 1691:                      Homo sapiens
#>                                                         Type
#>                                                       <char>
#>    1:     Expression profiling by high throughput sequencing
#>    2: Non-coding RNA profiling by high throughput sequencing
#>    3:     Expression profiling by high throughput sequencing
#>    4:                      Non-coding RNA profiling by array
#>    5:     Expression profiling by high throughput sequencing
#>   ---                                                       
#> 1687:                          Expression profiling by array
#> 1688:                          Expression profiling by array
#> 1689:                          Expression profiling by array
#> 1690:                     Protein profiling by protein array
#> 1691:                          Expression profiling by array
#>                                                                    FTP download
#>                                                                          <char>
#>    1:     GEO (XLSX) ftp://ftp.ncbi.nlm.nih.gov/geo/series/GSE277nnn/GSE277813/
#>    2:      GEO (TXT) ftp://ftp.ncbi.nlm.nih.gov/geo/series/GSE106nnn/GSE106837/
#>    3: GEO (MTX, TSV) ftp://ftp.ncbi.nlm.nih.gov/geo/series/GSE274nnn/GSE274264/
#>    4:      GEO (CEL) ftp://ftp.ncbi.nlm.nih.gov/geo/series/GSE283nnn/GSE283040/
#>    5: GEO (MTX, TSV) ftp://ftp.ncbi.nlm.nih.gov/geo/series/GSE276nnn/GSE276815/
#>   ---                                                                          
#> 1687:                GEO ftp://ftp.ncbi.nlm.nih.gov/geo/series/GSE2nnn/GSE2956/
#> 1688:                GEO ftp://ftp.ncbi.nlm.nih.gov/geo/series/GSE1nnn/GSE1322/
#> 1689:          GEO (CEL) ftp://ftp.ncbi.nlm.nih.gov/geo/series/GSE1nnn/GSE1009/
#> 1690:                  GEO ftp://ftp.ncbi.nlm.nih.gov/geo/series/GSEnnn/GSE634/
#> 1691:                  GEO ftp://ftp.ncbi.nlm.nih.gov/geo/series/GSEnnn/GSE121/
#>              ID                                           SRA Run Selector
#>           <int>                                                     <char>
#>    1: 200277813                                                       <NA>
#>    2: 200106837 https://www.ncbi.nlm.nih.gov/Traces/study/?acc=PRJNA418157
#>    3: 200274264                                                       <NA>
#>    4: 200283040                                                       <NA>
#>    5: 200276815                                                       <NA>
#>   ---                                                                     
#> 1687: 200002956                                                       <NA>
#> 1688: 200001322                                                       <NA>
#> 1689: 200001009                                                       <NA>
#> 1690: 200000634                                                       <NA>
#> 1691: 200000121                                                       <NA>
#>       Project                       Contains                           Datasets
#>        <char>                         <char>                             <char>
#>    1:    <NA>                     99 Samples                               <NA>
#>    2:    <NA>                    244 Samples                               <NA>
#>    3:    <NA>                     10 Samples                               <NA>
#>    4:    <NA>                     16 Samples                               <NA>
#>    5:    <NA>                      2 Samples                               <NA>
#>   ---                                                                          
#> 1687:    <NA>                       1 Sample                               <NA>
#> 1688:    <NA>                     21 Samples                               <NA>
#> 1689:    <NA>                      6 Samples                             GDS961
#> 1690:    <NA>                      4 Samples                               <NA>
#> 1691:    <NA> 5 related Platforms 50 Samples GDS157 GDS158 GDS160 GDS161 GDS162
#>               Platforms Series Accession
#>                  <char>           <char>
#>    1:          GPL24676        GSE277813
#>    2: GPL19057 GPL18573        GSE106837
#>    3:          GPL24676        GSE274264
#>    4:          GPL19117        GSE283040
#>    5:          GPL24676        GSE276815
#>   ---                                   
#> 1687:            GPL310          GSE2956
#> 1688:            GPL887          GSE1322
#> 1689:           GPL8300          GSE1009
#> 1690:            GPL120           GSE634
#> 1691:              <NA>           GSE121