Open the GEO landing page in a browser

Construct a GEO landing page and open it directly in the system's default web browser (or a user-specified browser).

Usage

geo_show(
  accession,
  famount = NULL,
  scope = NULL,
  over_https = NULL,
  browser = getOption("browser")
)

Arguments

accession

A character of GEO accession IDs. Examples:

DataSets (GDS): "GDS505", "GDS606", "GDS1234", "GDS9999", etc.

Series (GSE): "GSE2", "GSE22", "GSE100", "GSE2000", etc.

Platforms (GPL): "GPL96", "GPL570", "GPL10558", etc.

Samples (GSM): "GSM12345", "GSM67890", "GSM112233", etc.

famount

A character specifying file/amount type requested. GEO data can be accessed through two sites:

Direct FTP/HTTPS file retrieval from GEO FTP server (file type):

"soft": SOFT (Simple Omnibus in Text Format) from GEO FTP site. When accession is DataSets or Series, this is the default.

"soft_full": full SOFT (Simple Omnibus in Text Format) files from GEO FTP site by DataSet (GDS) containging additionally contains up-to-date gene annotation for the DataSet Platform.

"miniml": MINiML (MIAME Notation in Markup Language, pronounced miniml) is an XML format that incorporates experimental data and metadata. MINiML is essentially an XML rendering of SOFT format.

"matrix": Series matrix file.

"annot": annotation files for Platforms.

"suppl": supplementary files.

The following table summarizes the compatibility between GEO accession types and file type options:

famount	GDS	GSE	GPL	GSM

SOFT (soft)	o	o	o	x

SOFTFULL (soft_full)	o	x	x	x

MINiML (miniml)	x	o	o	x

Matrix (matrix)	x	o	x	x

Annotation (annot)	x	x	o	x

Supplementaryfiles (suppl)	x	o	o	o

For file retrieval from Accession Display Bar of GEO database (amount of data):

"none": Applicable only to DataSets; for DataSets, this is also the sole valid option.

"brief": accession attributes only.

"quick": accession attributes + first 20 rows of the data table.

"data": omits the accession's attributes, showing only links to other accessions and the full data table.

"full": accession attributes + complete data table.

scope

A character specifying which GEO accessions to include (Only applicable to Accession Display Bar access).

"none": Applicable only to DataSets; for DataSets, this is also the sole valid option

"self": the queried accession only.

"gsm", "gpl", "gse": related samples, platforms, or series.

"all": all accessions related to the query (family view).

over_https

Logical scalar. If TRUE, connects to GEO FTP server via HTTPS (https://ftp.ncbi.nlm.nih.gov/geo); otherwise uses plain FTP (ftp://ftp.ncbi.nlm.nih.gov/geo). Only applicable to GEO FTP server access.

browser

a non-empty character string giving the name of the program to be used as the HTML browser. It should be in the PATH, or a full path specified. Alternatively, an R function to be called to invoke the browser.

Under Windows NULL is also allowed (and is the default), and implies that the file association mechanism will be used.

Details

See utils::browseURL()